# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Platform API",
    "description": "Enterprise Healthcare Platform Foundation API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/fhir/Patient/{patient_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Patient Resource",
        "operationId": "get_fhir_patient_api_v1_fhir_Patient__patient_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Patient Api V1 Fhir Patient  Patient Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Practitioner/{practitioner_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Practitioner Resource",
        "operationId": "get_fhir_practitioner_api_v1_fhir_Practitioner__practitioner_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "practitioner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Practitioner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Practitioner Api V1 Fhir Practitioner  Practitioner Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Organization/{facility_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Organization Resource",
        "operationId": "get_fhir_organization_api_v1_fhir_Organization__facility_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Facility Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Organization Api V1 Fhir Organization  Facility Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Appointment/{appointment_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Appointment Resource",
        "operationId": "get_fhir_appointment_api_v1_fhir_Appointment__appointment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Appointment Api V1 Fhir Appointment  Appointment Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Encounter/{encounter_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Encounter Resource",
        "operationId": "get_fhir_encounter_api_v1_fhir_Encounter__encounter_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Encounter Api V1 Fhir Encounter  Encounter Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Observation/{observation_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Observation Resource (Vital Sign or Diagnostic Result)",
        "operationId": "get_fhir_observation_api_v1_fhir_Observation__observation_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Observation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Observation Api V1 Fhir Observation  Observation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/Medication/{medication_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 Medication Resource",
        "operationId": "get_fhir_medication_api_v1_fhir_Medication__medication_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "medication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Medication Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Medication Api V1 Fhir Medication  Medication Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/MedicationRequest/{prescription_item_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 MedicationRequest Resource",
        "operationId": "get_fhir_medication_request_api_v1_fhir_MedicationRequest__prescription_item_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Medication Request Api V1 Fhir Medicationrequest  Prescription Item Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/ServiceRequest/{service_request_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 ServiceRequest Resource (Diagnostic Order or Referral)",
        "operationId": "get_fhir_service_request_api_v1_fhir_ServiceRequest__service_request_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Service Request Api V1 Fhir Servicerequest  Service Request Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/DiagnosticReport/{diagnostic_order_id}": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Get FHIR R4 DiagnosticReport Resource",
        "operationId": "get_fhir_diagnostic_report_api_v1_fhir_DiagnosticReport__diagnostic_order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostic Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Diagnostic Report Api V1 Fhir Diagnosticreport  Diagnostic Order Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/patient/{patient_id}/bundle": {
      "get": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Export Complete Longitudinal Patient FHIR R4 Bundle",
        "operationId": "export_patient_bundle_api_v1_fhir_patient__patient_id__bundle_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "bundle_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(collection|document)$",
              "default": "collection",
              "title": "Bundle Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Patient Bundle Api V1 Fhir Patient  Patient Id  Bundle Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fhir/patient/{patient_id}/export-async": {
      "post": {
        "tags": [
          "FHIR R4 Interoperability"
        ],
        "summary": "Trigger Asynchronous Background FHIR Bundle Export (Celery)",
        "operationId": "export_patient_bundle_async_api_v1_fhir_patient__patient_id__export_async_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}